# backend/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing scheduled tasks and task groups",
        "title": "Cron Observer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/projects/{project_id}/task-groups": {
            "post": {
                "description": "Create a new task group in a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Create a new task group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task group creation request",
                        "name": "task_group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TaskGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{project_id}/task-groups/{group_uuid}": {
            "get": {
                "description": "Retrieve a task group by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Get a task group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Group UUID",
                        "name": "group_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing task group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Update a task group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Group UUID",
                        "name": "group_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task group update request",
                        "name": "task_group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTaskGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing task group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Delete a task group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Group UUID",
                        "name": "group_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{project_id}/task-groups/{group_uuid}/start": {
            "post": {
                "description": "Manually start all tasks in a task group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Start a task group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Group UUID",
                        "name": "group_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{project_id}/task-groups/{group_uuid}/stop": {
            "post": {
                "description": "Manually stop all tasks in a task group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Stop a task group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Group UUID",
                        "name": "group_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{project_id}/task-groups/{group_uuid}/tasks": {
            "get": {
                "description": "Retrieve all tasks belonging to a task group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-groups"
                ],
                "summary": "Get tasks in a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Group UUID",
                        "name": "group_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateTaskGroupRequest": {
            "type": "object",
            "required": [
                "name",
                "project_id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "end_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "project_id": {
                    "type": "string"
                },
                "start_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskGroupStatus"
                        }
                    ]
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "project_id is required"
                    ]
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "models.ScheduleType": {
            "type": "string",
            "enum": [
                "RECURRING",
                "ONEOFF"
            ],
            "x-enum-varnames": [
                "ScheduleTypeRecurring",
                "ScheduleTypeOneOff"
            ]
        },
        "models.Task": {
            "description": "Task represents a scheduled task entity",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Backup database daily"
                },
                "id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "example": "Daily Backup"
                },
                "project_id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "schedule_config": {
                    "$ref": "#/definitions/models.ScheduleConfig"
                },
                "schedule_type": {
                    "enum": [
                        "RECURRING",
                        "ONEOFF"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ScheduleType"
                        }
                    ],
                    "example": "RECURRING"
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "ACTIVE"
                },
                "task_group_id": {
                    "description": "Optional reference to task group",
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "trigger_config": {
                    "$ref": "#/definitions/models.TriggerConfig"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "models.TaskGroup": {
            "description": "TaskGroup represents a group of tasks that can be controlled together",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Tasks that run in the morning"
                },
                "end_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string",
                    "example": "17:00"
                },
                "id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "name": {
                    "type": "string",
                    "example": "Morning Tasks"
                },
                "project_id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "start_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string",
                    "example": "09:00"
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskGroupStatus"
                        }
                    ],
                    "example": "ACTIVE"
                },
                "timezone": {
                    "description": "IANA timezone (e.g., \"America/New_York\")",
                    "type": "string",
                    "example": "America/New_York"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "models.TaskGroupStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "PAUSED",
                "DISABLED"
            ],
            "x-enum-varnames": [
                "TaskGroupStatusActive",
                "TaskGroupStatusPaused",
                "TaskGroupStatusDisabled"
            ]
        },
        "models.TaskStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "PAUSED",
                "DISABLED"
            ],
            "x-enum-varnames": [
                "TaskStatusActive",
                "TaskStatusPaused",
                "TaskStatusDisabled"
            ]
        },
...
        "models.UpdateTaskGroupRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "end_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "start_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskGroupStatus"
                        }
                    ]
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
    }
}